+++ packages/nyris-webapp/src/components/SidePanel.tsx
@@ -2,11 +2,14 @@ import { Box, Button, Typography } from '@material-ui/core';
 import { RectCoords } from '@nyris/nyris-api';
 import { Preview } from '@nyris/nyris-react-components';
 import React, { useState } from 'react';
-import ExpandablePanelComponent from './PanelResult';
+import PostFilterPanel from './PanelResult/PostFilter';
+import PostFilterPanelAlgolia from './PanelResult/PostFilterAlgolia';
+
 import { useTranslation } from 'react-i18next';
 import { useAppSelector } from 'Store/Store';
-import KeyboardArrowRightOutlinedIcon from '@material-ui/icons/KeyboardArrowRightOutlined';
-import KeyboardArrowLeftOutlinedIcon from '@material-ui/icons/KeyboardArrowLeftOutlined';
+import { ReactComponent as KeyboardArrowRightOutlinedIcon } from 'common/assets/icons/arrow_right.svg';
+import { ReactComponent as KeyboardArrowLeftOutlinedIcon } from 'common/assets/icons/arrow_left.svg';
+
 import { DEFAULT_REGION } from '../constants';
 import { ReactComponent as IconInfo } from 'common/assets/icons/info-tooltip.svg';
 
@@ -32,7 +35,7 @@ function SidePanel({
 }) {
   const { t } = useTranslation();
   const [toggleColLeft, setToggleColLeft] = useState<boolean>(false);
-  const stateGlobal = useAppSelector((state: any) => state);
+  const stateGlobal = useAppSelector(state => state);
   const { search, settings } = stateGlobal;
 
   const { requestImage } = search;
@@ -49,7 +52,7 @@ function SidePanel({
       <Box
         className="box-toggle-coloumn"
         style={{
-          right: requestImage || toggleColLeft ? '0px' : '16px',
+          right: '0px',
         }}
       >
         <Button
@@ -142,8 +145,16 @@ function SidePanel({
         )}
 
         {showPostFilter && (
-          <Box className="col-left__bottom">
-            <ExpandablePanelComponent disjunctiveFacets={disjunctiveFacets} />
+          <Box
+            className="col-left__bottom"
+            style={{
+              marginTop: requestImage ? '16px' : '48px',
+            }}
+          >
+            {settings.algolia.enabled && (
+              <PostFilterPanelAlgolia disjunctiveFacets={disjunctiveFacets} />
+            )}
+            {!settings.algolia.enabled && <PostFilterPanel />}
           </Box>
         )}
       </Box>