+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.3.48 - 2024-03-01
+
+### Added
+- Widget release
+- DE translations for pre-filter
+- Updated example settings file
+- Added post filter for Elastic version of search suite
+
+### Fix
+- On enter button the page refreshes
+
 ## 0.3.47 - 2024-01-31
 
 ### Added